refactor(fs): 删掉检查器数据库死代码，拆开 browser
去掉 embed 第二套浏览器、recordFocus/inspectorHost，以及 inspector-database 相关文件。单元格、样式和记录详情从 browser.tsx 拆出。
