+++ packages/core-file-system/src/web/database-ui.test.ts
@@ -25,25 +25,18 @@ test('decorate merges cells; later layer wins; dispose restores', () => {
   assert.equal(ui.chrome('/plugins').cells?.name, undefined)
 })
 
-test('record focus and inspector host notify subscribers', () => {
+test('decorate notifies subscribers', () => {
   const ctx = new Context()
   const ui = new DatabaseUiService(ctx)
   let n = 0
   const off = ui.subscribe(() => {
     n += 1
   })
-  ui.setRecordFocus({ path: '/pages', recordId: 'r1' })
-  assert.deepEqual(ui.getRecordFocus(), { path: '/pages', recordId: 'r1' })
-  const host = { tagName: 'DIV' } as HTMLElement
-  ui.setInspectorHost(host)
-  assert.equal(ui.getInspectorHost(), host)
-  ui.setRecordFocus({ path: '/pages', recordId: 'r1' })
-  ui.setInspectorHost(host)
+  const handle = ui.decorate('/pages', { cells: {} })
+  assert.equal(n, 1)
+  handle.dispose()
   assert.equal(n, 2)
-  ui.setRecordFocus(null)
-  ui.setInspectorHost(null)
   off()
-  assert.equal(ui.getRecordFocus(), null)
 })
 
 test('decorate panes with the same id keep a single pane', () => {
