+++ packages/core-file-system/src/web/sidebar-nav.ts
@@ -1,6 +1,6 @@
 /** 侧栏展开/收起与数据路由切换的纯逻辑，UI 和压测共用。 */
 import { parseAppPath } from '@biu/web-session-view'
-import { databaseRecordPath, databaseViewPath } from './inspector-nav.ts'
+import { databaseRecordPath, databaseViewPath } from './database-path.ts'
 import type { ViewMode } from './fields.ts'
 
 const DATABASE = [{ id: 'database', label: '数据', path: '/database' }]
